Stability – Miscellaneous code cleanups and fixes: - Fixed some threading problems in File Util. - Fixed Tile Service unit tests (there were some codes previously commented out due to failures). - Got rid of some compiling warnings in FDO Connection Manager. - Implemented tile file locking mechanism to address some threading/file access problems.
git-svn-id: http://svn.osgeo.org/mapguide/trunk/MgDev@1090 a442ad82-b40c-0410-ae44-e6d7bd6e4a36

+++ Server/src/Common/Manager/FdoConnectionManager.cpp
@@ -1234,7 +1234,8 @@ MdfModel::FeatureSource* MgFdoConnectionManager::GetFeatureSource(MgResourceIden
     RetrieveFeatureSource(resId, featureSourceXmlContent);
 
     MdfParser::FSDSAX2Parser parser;
-    parser.ParseString(featureSourceXmlContent.c_str(), featureSourceXmlContent.length()*sizeof(char));
+    parser.ParseString(featureSourceXmlContent.c_str(),
+        (unsigned int)(featureSourceXmlContent.length()*sizeof(char)));
 
     assert(parser.GetSucceeded());
